Hard-code the HTML for the 2-factor webauth error
We used the GOVUK Frontend error message component
for this. Recent updates to our version of GOVUK
Frontend changed the elements used for this
message from `<span>`s to `<p>`s. This broke our
code, meaning part of it is now showing.

This hard-codes the original HTML, until we can
fix this properly (in
https://trello.com/c/byv6y7Wa/133-fix-webauth-error-messages).

This problem also appeared on the page where you
authenticate with your 2-factor thing:

https://github.com/alphagov/notifications-admin/pull/4469/

At the time I thought this didn't effect the
security keys page on your profile. This proves
that was wrong.