move letter_branding flow to its own file
not because it's immensley complex right now, but more to provide an
example for other groups of endpoints.

the platform admin views are currently still in index.py as they're
deduped across letters and emails. we might want to pull them out later
just to remove all references to any kind of branding from
service_settings but im not fussed about that now.